requirements.txt
Flask==2.3.0
mysql-connector-python==8.0.33
pytest==8.3.3
pytest-mock==3.14.0

